Refactor Quota-related data classes to its own QuotaModel file in line with our current structure. Also change data type of prices from int to double for Market-related data classes as it is more appropriate.
