+++ app/src/main/java/com/example/farmeraid/data/MarketRepository.kt
@@ -39,7 +39,7 @@ class MarketRepository(
             MarketModel.Market(
                 id = market.id,
                 name = market.get("name") as String,
-                prices = market.get("prices") as MutableMap<String, Int>
+                prices = market.get("prices") as MutableMap<String, Double>
             )
                 .let { marketModelList.add(it) }
         }
@@ -72,7 +72,7 @@ class MarketRepository(
                             id = market.id,
                             name = market.get("name") as String,
                             quota = it,
-                            prices = market.get("prices") as MutableMap<String, Int>
+                            prices = market.get("prices") as MutableMap<String, Double>
                         )
                     )
                 } ?: run {
@@ -119,7 +119,7 @@ class MarketRepository(
                         id = marketModel.id,
                         name = marketModel.get("name") as String,
                         quota = it,
-                        prices = marketModel.get("prices") as MutableMap<String, Int>
+                        prices = marketModel.get("prices") as MutableMap<String, Double>
                     )
                 } ?: run {
                     return ResponseModel.FAResponseWithData.Error(quotaResponse.error ?: "Unknown error while fetching market's quota")
